core-lib: Implement UpdateOp::update according to a TODO in view.rs.
(fixes #1301)
(fixes #1228)

This change introduces a new UpdateOp, "update", in addition to
"copy", "insert", "invalidate" and "skip".

UpdateOp::update is sent by View::send_update_for_plan() in case of an
update of a line cache shadow segment if this update only moves the
cursor (!CURSOR_VALID) and leaves the text and styles intact
(TEXT_VALID && STYLES_VALID).

As a side effect, this change fixes the problem with an "invalidate" op
showing up in the middle of the vector of ops. The assertion failure in
test_invalidate() is now gone.

It also fixes a couple minor bugs in line/region overlaps calculation in
View::render_line().

This change requires the frontends to implement the support for the
new notification type (if not done already).

Because of the incompatible change in behavior, xi-core version is
bumped to 0.4.